Fold tests/test-gzip-error.js into tests/test-gzip.js
